requirements.txt
fastapi
uvicorn
python-dateutil
pydantic
pytest
httpx
python-dotenv